+++ packages/server/dataloader/__tests__/isCompanyDomain.test.ts
@@ -1,5 +1,5 @@
 import '../../../../scripts/webpack/utils/dotenv'
-import {getNewDataLoader} from '../../graphql/getDataLoader'
+import {getNewDataLoader} from '../getNewDataLoader'
 
 const dataloader = getNewDataLoader()
