+++ packages/server/dataloader/getInMemoryDataLoader.ts
@@ -1,0 +1,5 @@
+import {dataLoaderCache} from './RootDataLoader'
+
+export const getInMemoryDataLoader = (id: string) => {
+  return dataLoaderCache.use(id)
+}
