+++ packages/server/dataloader/getRedisDataLoader.ts
@@ -1,0 +1,10 @@
+import getRedis from '../utils/getRedis'
+import {hydrateDataLoader} from './hydrateDataLoader'
+
+export const getRedisDataLoader = async (id: string) => {
+  const serializedDataLoader = await getRedis().get(`dataLoader:${id}`)
+  if (serializedDataLoader) {
+    return hydrateDataLoader(id, serializedDataLoader)
+  }
+  return undefined
+}
