+++ packages/server/dataloader/githubLoaders.ts
@@ -5,7 +5,7 @@ import getGitHubAuthByUserIdTeamId, {
 import getGitHubDimensionFieldMaps, {
   GitHubDimensionFieldMap
 } from '../postgres/queries/getGitHubDimensionFieldMaps'
-import RootDataLoader from './RootDataLoader'
+import type RootDataLoader from './RootDataLoader'
 
 export const githubAuth = (parent: RootDataLoader) => {
   return new DataLoader<{teamId: string; userId: string}, GitHubAuth | null, string>(
