+++ packages/server/graphql/public/mutations/helpers/__tests__/getIsEmailApprovedByOrg.test.ts
@@ -1,7 +1,7 @@
 import '../../../../../../../scripts/webpack/utils/dotenv'
 import {createPGTables} from '../../../../../__tests__/common'
+import {getNewDataLoader} from '../../../../../dataloader/getNewDataLoader'
 import getKysely from '../../../../../postgres/getKysely'
-import {getNewDataLoader} from '../../../../getDataLoader'
 import getIsEmailApprovedByOrg from '../getIsEmailApprovedByOrg'
 
 const TEST_DB = 'getIsEmailApprovedByOrgTest'
