+++ packages/server/wsHandler.ts
@@ -9,7 +9,7 @@ import {Threshold} from '../client/types/constEnums'
 import sleep from '../client/utils/sleep'
 import activeClients from './activeClients'
 import AuthToken from './database/types/AuthToken'
-import {getNewDataLoader} from './graphql/getDataLoader'
+import {getNewDataLoader} from './dataloader/getNewDataLoader'
 import getRateLimiter from './graphql/getRateLimiter'
 import privateSchema from './graphql/private/rootSchema'
 import checkBlacklistJWT from './utils/checkBlacklistJWT'
